Ready for step 1 RELEASE

+++ src/main/kotlin/lt/skafis/bankas/service/implementations/ApprovalServiceImpl.kt
@@ -102,7 +102,7 @@ class ApprovalServiceImpl : ApprovalService {
         if (source.reviewStatus == ReviewStatus.APPROVED) {
             val sourceProblems = problemRepository.getBySourceId(sourceId)
             sourceProblems.forEach {
-                problemRepository.update(it.copy(skfCode = "", isApproved = false), it.id)
+                problemRepository.update(it.copy(isApproved = false), it.id)
                 if (it.isApproved) {
                     metaService.removeSkfCodeFromUsedList(it.skfCode)
                 }
@@ -250,7 +250,9 @@ class ApprovalServiceImpl : ApprovalService {
             throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
         }
         problemRepository.delete(problemId)
-        metaService.removeSkfCodeFromUsedList(problem.skfCode)
+        if (problem.skfCode.isNotEmpty()) {
+            metaService.removeSkfCodeFromUsedList(problem.skfCode)
+        }
         if (problem.problemImagePath.startsWith("problems/")) {
             storageRepository.deleteImage(problem.problemImagePath)
         }
